requirements.txt
Flask==0.12.2
transitions==0.6.3
pygraphviz==1.3.1
python-telegram-bot==5.3.0
selenium==3.8.0
